requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
typing
dataclasses
PyYAML
duckduckgo-search
feedparser
pandas
Pillow
python-docx
PyPDF2
openpyxl
tooluniverse
setuptools
neo4j
